requirements.txt
PyMuPDF
PyExifTool==0.4.13
Pillow
pyail
pdfx
